refactor : StepStatus -> PathStatus 이름 변경
