+++ src/main/java/ladder/dto/LineResult.java
@@ -1,11 +1,11 @@
 package ladder.dto;
 
 import ladder.domain.Line;
-import ladder.domain.StepStatus;
+import ladder.domain.PathStatus;
 
 import java.util.List;
 
-public record LineResult(List<StepStatus> value) {
+public record LineResult(List<PathStatus> value) {
 
     public static LineResult of(Line line) {
         return new LineResult(line.getStepStatuses());
