+++ src/test/java/ladder/domain/LinesTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static ladder.domain.StepStatus.EXIST;
-import static ladder.domain.StepStatus.NONE;
+import static ladder.domain.PathStatus.EXIST;
+import static ladder.domain.PathStatus.NONE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class LinesTest {
