+++ src/test/java/ladder/domain/PathStatusTest.java
@@ -3,15 +3,15 @@ package ladder.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static ladder.domain.StepStatus.EXIST;
+import static ladder.domain.PathStatus.EXIST;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class StepStatusTest {
+public class PathStatusTest {
 
     @DisplayName("true가 입력되면 EXIST를 반환한다.")
     @Test
     void returnNoneWhenStepExist() {
-        assertThat(StepStatus.getStepStatus(true))
+        assertThat(PathStatus.getStepStatus(true))
                 .isEqualTo(EXIST);
     }
